refactor(capabilities): extract shared fresh_substrate test helper (issue 786)
The same 'fn fresh_substrate() -> (Arc<Registry>, Arc<Mailer>)' was
copy-pasted across six aether-capabilities test modules — three with
a loopback HubOutbound wired (engine/proxy, rpc/client, rpc/server)
and three without (audio, fs, http). Collapse to one canonical
'pub(crate) fn fresh_substrate' in the test_chassis module that
ships the loopback-attached variant; the unknown-mailbox bubble-up
path stays inert for the three cap-internal sites (rx dropped, the
audio/fs/http tests never hit it) and remains live for the three
rpc/proxy sites that already depended on it.

Net -57 lines across the six call sites.

Closes iamacoffeepot/aether#786
