+++ crates/aether-capabilities/src/engine/proxy.rs
@@ -444,30 +444,15 @@ mod tests {
     use crate::rpc::server::{RpcServerCapability, RpcServerConfig, RpcServerHandle};
     use crate::rpc::test_echo::{TestEchoActor, TestEchoRequest};
     use crate::rpc::wire::PeerKind;
-    use crate::test_chassis::TestChassis;
+    use crate::test_chassis::{TestChassis, fresh_substrate};
     use crate::trace::TraceObserverCapability;
     use aether_actor::Actor;
     use aether_data::{EngineId, Kind, Uuid, mailbox_id_from_name};
     use aether_substrate::Subname;
     use aether_substrate::chassis::builder::Builder;
-    use aether_substrate::handle_store::HandleStore;
-    use aether_substrate::mail::mailer::Mailer;
-    use aether_substrate::mail::outbound::HubOutbound;
-    use aether_substrate::mail::registry::Registry;
     use aether_substrate::mail::{Mail, ReplyTarget, ReplyTo};
     use std::sync::{Arc, Mutex};
     use std::time::{Duration, Instant};
-
-    fn fresh_substrate() -> (Arc<Registry>, Arc<Mailer>) {
-        let registry = Arc::new(Registry::new());
-        for d in aether_kinds::descriptors::all() {
-            let _ = registry.register_kind_with_descriptor(d);
-        }
-        let (outbound, _rx) = HubOutbound::attached_loopback();
-        let store = Arc::new(HandleStore::new(1024 * 1024));
-        let mailer = Arc::new(Mailer::new(Arc::clone(&registry), store).with_outbound(outbound));
-        (registry, mailer)
-    }
 
     fn substrate_peer_kind() -> PeerKind {
         PeerKind::Substrate {
