+++ crates/aether-capabilities/src/http.rs
@@ -456,21 +456,8 @@ mod native {
         use aether_substrate::chassis::error::BootError;
         use aether_substrate::mail::ReplyTo;
         use aether_substrate::mail::mailer::Mailer;
-        use aether_substrate::mail::registry::Registry;
 
-        use crate::test_chassis::TestChassis;
-
-        fn fresh_substrate() -> (Arc<Registry>, Arc<Mailer>) {
-            let registry = Arc::new(Registry::new());
-            for d in aether_kinds::descriptors::all() {
-                let _ = registry.register_kind_with_descriptor(d);
-            }
-            let store = ::std::sync::Arc::new(::aether_substrate::handle_store::HandleStore::new(
-                1024 * 1024,
-            ));
-            let mailer = Arc::new(Mailer::new(Arc::clone(&registry), store));
-            (registry, mailer)
-        }
+        use crate::test_chassis::{TestChassis, fresh_substrate};
 
         struct StubAdapter {
             response: Mutex<Option<Result<FetchResponse, HttpError>>>,
