+++ crates/aether-capabilities/src/rpc/client.rs
@@ -271,33 +271,18 @@ mod tests {
     use crate::rpc::wire::{
         HelloAck, MailEnvelope, MailboxAddress, PeerKind, WIRE_VERSION, WireFrame,
     };
-    use crate::test_chassis::TestChassis;
+    use crate::test_chassis::{TestChassis, fresh_substrate};
     use crate::trace::TraceObserverCapability;
     use aether_actor::Actor;
     use aether_codec::frame::{read_frame, write_frame};
     use aether_data::{Kind, mailbox_id_from_name};
     use aether_substrate::chassis::builder::Builder;
-    use aether_substrate::handle_store::HandleStore;
-    use aether_substrate::mail::mailer::Mailer;
-    use aether_substrate::mail::outbound::HubOutbound;
-    use aether_substrate::mail::registry::Registry;
     use std::io::BufReader;
     use std::net::{TcpListener, TcpStream};
     use std::sync::Arc;
     use std::sync::atomic::{AtomicUsize, Ordering};
     use std::thread::JoinHandle;
     use std::time::Duration;
-
-    fn fresh_substrate() -> (Arc<Registry>, Arc<Mailer>) {
-        let registry = Arc::new(Registry::new());
-        for d in aether_kinds::descriptors::all() {
-            let _ = registry.register_kind_with_descriptor(d);
-        }
-        let (outbound, _rx) = HubOutbound::attached_loopback();
-        let store = Arc::new(HandleStore::new(1024 * 1024));
-        let mailer = Arc::new(Mailer::new(Arc::clone(&registry), store).with_outbound(outbound));
-        (registry, mailer)
-    }
 
     fn substrate_peer_kind() -> PeerKind {
         PeerKind::Substrate {
